requirements.txt
asgiref==3.2.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1
django-allauth==0.41.0
django-toolbelt==0.0.1
djangorestframework==3.11.0
elasticsearch==7.0.0
google-api-core==1.15.0
google-auth==1.10.0
google-cloud==0.34.0
google-cloud-core==1.1.0
google-cloud-storage==1.24.1
google-cloud-vision==0.41.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.26.0
gunicorn==20.0.4
idna==2.8
Jinja2==2.10.3
jsonfield==2.0.2
MarkupSafe==1.1.1
numpy==1.18.0
oauthlib==3.1.0
pandas==0.25.3
Pillow==7.0.0
protobuf==3.11.2
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyMySQL==0.9.3
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.7
